Fix "cast from size_t to int" warnings.

+++ external/utf8conv/utf8conv_inl.h
@@ -87,7 +87,7 @@ inline std::wstring UTF16FromUTF8(const std::string & utf8)
         CP_UTF8,            // convert from UTF-8
         conversionFlags,    // flags
         utf8.data(),        // source UTF-8 string
-        utf8.length(),      // length (in chars) of source UTF-8 string
+        static_cast<int>(utf8.length()),      // length (in chars) of source UTF-8 string
         NULL,               // unused - no conversion done in this step
         0                   // request size of destination buffer, in wchar_t's
         );
@@ -120,9 +120,9 @@ inline std::wstring UTF16FromUTF8(const std::string & utf8)
         0,                  // validation was done in previous call, 
                             // so speed up things with default flags
         utf8.data(),        // source UTF-8 string
-        utf8.length(),      // length (in chars) of source UTF-8 string
+        static_cast<int>(utf8.length()),      // length (in chars) of source UTF-8 string
         &utf16[0],          // destination buffer
-        utf16.length()      // size of destination buffer, in wchar_t's
+        static_cast<int>(utf16.length())      // size of destination buffer, in wchar_t's
         ) )
     {
         // Error
@@ -164,7 +164,7 @@ inline std::wstring UTF16FromUTF8(const char * utf8, size_t uBufferSize)
         CP_UTF8,            // convert from UTF-8
         conversionFlags,    // flags
         utf8,               // source UTF-8 string
-        utf8Length,         // length (in chars) of source UTF-8 string
+        static_cast<int>(utf8Length),         // length (in chars) of source UTF-8 string
         NULL,               // unused - no conversion done in this step
         0                   // request size of destination buffer, in wchar_t's
         );
@@ -196,9 +196,9 @@ inline std::wstring UTF16FromUTF8(const char * utf8, size_t uBufferSize)
         0,                  // validation was done in previous call, 
                             // so speed up things with default flags
         utf8,               // source UTF-8 string
-        utf8Length,         // length (in chars) of source UTF-8 string
+        static_cast<int>(utf8Length),         // length (in chars) of source UTF-8 string
         &utf16[0],          // destination buffer
-        utf16.length()      // size of destination buffer, in wchar_t's
+        static_cast<int>(utf16.length())      // size of destination buffer, in wchar_t's
         ) )
     {
         // Error
@@ -234,7 +234,7 @@ inline std::string UTF8FromUTF16(const std::wstring & utf16)
         CP_UTF8,            // convert to UTF-8
         0,                  // default flags
         utf16.data(),       // source UTF-16 string
-        utf16.length(),     // source string length, in wchar_t's,
+        static_cast<int>(utf16.length()),     // source string length, in wchar_t's,
         NULL,               // unused - no conversion required in this step
         0,                  // request buffer size
         NULL, NULL          // unused
@@ -264,9 +264,9 @@ inline std::string UTF8FromUTF16(const std::wstring & utf16)
         CP_UTF8,                // convert to UTF-8
         0,                      // default flags
         utf16.data(),           // source UTF-16 string
-        utf16.length(),         // source string length, in wchar_t's,
+        static_cast<int>(utf16.length()),         // source string length, in wchar_t's,
         &utf8[0],               // destination buffer
-        utf8.length(),          // destination buffer size, in chars
+        static_cast<int>(utf8.length()),          // destination buffer size, in chars
         NULL, NULL              // unused
         ) )
     {
@@ -306,7 +306,7 @@ inline std::string UTF8FromUTF16(const wchar_t * utf16)
         CP_UTF8,            // convert to UTF-8
         0,                  // default flags
         utf16,              // source UTF-16 string
-        utf16Length,        // source string length, in wchar_t's,
+        static_cast<int>(utf16Length),        // source string length, in wchar_t's,
         NULL,               // unused - no conversion required in this step
         0,                  // request buffer size
         NULL, NULL          // unused
@@ -336,9 +336,9 @@ inline std::string UTF8FromUTF16(const wchar_t * utf16)
         CP_UTF8,                // convert to UTF-8
         0,                      // default flags
         utf16,                  // source UTF-16 string
-        utf16Length,            // source string length, in wchar_t's,
+        static_cast<int>(utf16Length),            // source string length, in wchar_t's,
         &utf8[0],               // destination buffer
-        utf8.length(),          // destination buffer size, in chars
+        static_cast<int>(utf8.length()),          // destination buffer size, in chars
         NULL, NULL              // unused
         ) )
     {
